% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{clean_numeric_iqr}
\alias{clean_numeric_iqr}
\title{Limpieza segun recorrido intercuantilico}
\usage{
clean_numeric_iqr(x, k = 1.5, qmin = 0.25, qmax = 0.75, ...)
}
\arguments{
\item{x}{vector a limpiar}

\item{k}{factor que determina la amplitud de la limpieza (k veces
el recorrido intercuantilico)}

\item{qmin}{cuantil inferior a utilizar}

\item{qmax}{cuantil superior a utilizar}
}
\value{
x
}
\description{
Limpieza segun recorrido intercuantilico
}
\examples{
clean_numeric_iqr(rnorm(0,50))
}
