% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifica_filas.R
\name{cuenta_repeticiones_consecutivas}
\alias{cuenta_repeticiones_consecutivas}
\title{Cuenta repeticiones y extrae valores repetido a lo largo de un vector}
\usage{
cuenta_repeticiones_consecutivas(x)
}
\arguments{
\item{x}{}
}
\description{
Cuenta repeticiones y extrae valores repetido a lo largo de un vector
}
