% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifica_filas.R
\name{identifica_filas}
\alias{identifica_filas}
\title{Con esta función se pretende agrupar con un indice tramos que compartan alguna característica en común}
\usage{
identifica_filas(x, condition, fill = NA)
}
\arguments{
\item{x}{}

\item{condition}{}

\item{fill}{}
}
\description{
Con esta función se pretende agrupar con un indice tramos que compartan alguna característica en común
}
