% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifica_filas.R
\name{ifilas_tramos}
\alias{ifilas_tramos}
\title{Funcion que identifica tramos con un incremento constante dado}
\usage{
ifilas_tramos(x, incremento)
}
\arguments{
\item{x}{}

\item{incremento}{}
}
\description{
Funcion que identifica tramos con un incremento constante dado
}
