% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifica_filas.R
\name{ifilas_tramos_tiempo_sup}
\alias{ifilas_tramos_tiempo_sup}
\title{Funcion que identifica tramos con un incremento no superior a un tiempo en segundos}
\usage{
ifilas_tramos_tiempo_sup(x, incremento)
}
\arguments{
\item{x}{}

\item{incremento}{}
}
\description{
Funcion que identifica tramos con un incremento no superior a un tiempo en segundos
}
