% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_templates.R
\name{izertis_report}
\alias{izertis_report}
\title{Creacion de template html de Izertis}
\usage{
izertis_report(toc = TRUE, toc_float = FALSE, keep_md = FALSE)
}
\arguments{
\item{toc}{}

\item{toc_float}{}
}
\description{
Creacion de template html de Izertis
}
