% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_izertis.R
\name{izertisPalette}
\alias{izertisPalette}
\title{Izertis theme palette}
\usage{
izertisPalette(n)
}
\arguments{
\item{n}{the number of colours to output}
}
\description{
This function outputs n colours from the fira ggplot2 theme palette
}
