% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_templates.R
\name{redes_report}
\alias{redes_report}
\title{Creacion de template html de Redes}
\usage{
redes_report(toc = TRUE, toc_float = FALSE, keep_md = FALSE)
}
\arguments{
\item{toc}{}

\item{toc_float}{}
}
\value{

}
\description{
Creacion de template html de Redes
}
