% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/izertis_theme.R
\name{theme_izertis}
\alias{theme_izertis}
\title{Izertis theme}
\usage{
theme_izertis(main_color1 = coral_red, main_color2 = cobalt,
  secondary_color1 = cream_can, secondary_color2 = mine_shaft)
}
\arguments{
\item{main_color1}{color principal 1}

\item{main_color2}{color principal 2}

\item{secondary_color1}{color secondary 1}

\item{secondary_color1}{color secondary 2}
}
\value{
ggplot theme
}
\description{
Izertis theme
}
\examples{
library(ggplot2)

ggplot(mtcars, aes(x = mpg*0.43, y = wt*0.4535924, colour = factor(cyl))) +
  geom_point(size = 2) +
  labs(title = "Car weight vs efficiency",
       subtitle = "Using sensible metrics",
       x = "Efficiency (km/l)",
       y = "Weight (1000 kg)",
       colour = "Cylinders") +
  theme_izertis() 

}
